backend/requirements.txt
flask
flask-session
mysql-connector-python
python-dotenv
werkzeug
